Everything Updated, Don't Modify

+++ Interview_Tracking_System/src/main/java/com/spring/service/AdminServiceImpl.java
@@ -30,7 +30,6 @@ import com.spring.utils.CandidateUtils;
 import com.spring.utils.HrPanelUtils;
 import com.spring.utils.InterviewScheduleUtils;
 import com.spring.utils.TechPanelUtils;
-import com.spring.utils.UserProfileUtils;
 @Service
 public class AdminServiceImpl implements AdminService {
 	
@@ -382,10 +381,10 @@ List<ITS_TBL_User_Credentials_Entity> userList=userRepository.findBysessionId(au
 	public Object sendToHr(long interviewId,String date,String time,long empHRId,String authToken){
 		List<ITS_TBL_User_Credentials_Entity> userList=userRepository.findBysessionId(authToken);
 		ITS_TBL_Interview_Schedule_Entity newInterview= schdeduleRepository.findByInterviewId(interviewId);
-		int checkTechclear=newInterview.getTechRating();
+		float checkTechclear=newInterview.getTechRating();
 		if(userList!=null && userList.size()!=0 && newInterview!=null)
 		{
-		if(checkTechclear>3)
+		if(checkTechclear>3f)
 		{
 		DateTimeFormatter newPattern = DateTimeFormatter.ofPattern("yyyy-MM-dd");
 		LocalDate localDate=LocalDate.parse(date,newPattern);
